requirements.txt
mysql-connector-python==8.0.29
tabulate==0.8.9
requests==2.28.0